+++ src/bson/bson-oid.h
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "bson/bson-prelude.h"
+#include "bson-prelude.h"
 
 
 #ifndef BSON_OID_H
@@ -23,10 +23,10 @@
 
 #include <time.h>
 
-#include "bson/bson-context.h"
-#include "bson/bson-macros.h"
-#include "bson/bson-types.h"
-#include "bson/bson-endian.h"
+#include "bson-context.h"
+#include "bson-macros.h"
+#include "bson-types.h"
+#include "bson-endian.h"
 
 
 BSON_BEGIN_DECLS
@@ -214,8 +214,8 @@ bson_oid_init_from_string_unsafe (bson_oid_t *oid, const char *str)
    int i;
 
    for (i = 0; i < 12; i++) {
-      oid->bytes[i] = ((bson_oid_parse_hex_char (str[2 * i]) << 4) |
-                       (bson_oid_parse_hex_char (str[2 * i + 1])));
+      oid->bytes[i] = (uint8_t) ((bson_oid_parse_hex_char (str[2 * i]) << 4) |
+                                 (bson_oid_parse_hex_char (str[2 * i + 1])));
    }
 }
